test-harness/requirements.txt
typer[all]>=0.9.0
rich>=13.0.0
gspread>=5.12.0
google-auth>=2.23.0
anthropic>=0.39.0
openai>=1.50.0
google-genai>=1.0.0
python-dotenv>=1.0.0
